Extract alarm to seperate module
