+++ assets/css/style.css
@@ -1,5 +1,3 @@
-@import 'alarm.css';
-
 body {
     font-family: Numans;
     font-size: 10px;
